mcp_server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "FastMCP-based MCP server built with FastMCP"
readme = "README.md"
requires-python = ">=3.12, <3.14"
dependencies = [
  "fastmcp==3.1.0",
  "httpx>=0.28.1",
  "pytest",
  "python-dotenv>=1.2.1",
]

